New logic in FooterInfo: contacts of each author are in array of objects. So now it's possible to have different contacts (not only on GitHub and LinkedIn) and more than 2 contacts.
