+++ src/components/Footer/Footer.js
@@ -4,81 +4,7 @@ import './Footer.scss';
 import FooterItem from './FooterItem/FooterItem';
 import FooterInfo from './FooterInfo/FooterInfo';
 
-import author from './../../img/Footer/author.png';
-import uszakow from '../../img/Footer/uszakow.png';
-
-const authors = [
-    {
-        name: 'Paweł Uszakow',
-        role: 'front-end developer',
-        photo: uszakow,
-        github: 'https://github.com/uszakow',
-        linkedin: 'https://www.linkedin.com/in/p-uszakow/'
-    },
-    {
-        name: '2Długieimię Długienazwisko',
-        role: 'funkcja rola',
-        photo: author,
-        github: '#',
-        linkedin: '#'
-    },
-    {
-        name: '3Długieimię Długienazwisko',
-        role: 'funkcja',
-        photo: author,
-        github: '#',
-        linkedin: '#'
-    },
-    {
-        name: '4Imię Nazwisko',
-        role: 'funkcja',
-        photo: author,
-        github: '#',
-        linkedin: '#'
-    },
-    {
-        name: '5Imię Nazwisko',
-        role: 'funkcja',
-        photo: author,
-        github: '#',
-        linkedin: '#'
-    },
-    {
-        name: '6Imię Nazwisko',
-        role: 'funkcja',
-        photo: author,
-        github: '#',
-        linkedin: '#'
-    },
-    {
-        name: '7Imię Nazwisko',
-        role: 'funkcja w projekcie',
-        photo: author,
-        github: '#',
-        linkedin: '#'
-    },
-    {
-        name: '8Długieimię Długienazwisko',
-        role: 'funkcja rola',
-        photo: author,
-        github: '#',
-        linkedin: '#'
-    },
-    {
-        name: '9Długieimię Długienazwisko',
-        role: 'funkcja rola',
-        photo: author,
-        github: '#',
-        linkedin: '#'
-    },
-    {
-        name: '10Imię Nazwisko',
-        role: 'funkcja w projekcie',
-        photo: author,
-        github: '#',
-        linkedin: '#'
-    },
-];
+import authors from './authors';
 
 class Footer extends Component {
     constructor(props) {
