+++ src/components/Footer/authors.js
@@ -1,0 +1,98 @@
+import author from './../../img/Footer/author.png';
+import uszakow from '../../img/Footer/uszakow.png';
+
+const authors = [
+    {
+        name: 'Paweł Uszakow',
+        role: 'front-end developer',
+        photo: uszakow,
+        contacts: [
+            { "GitHub": "https://github.com/uszakow" },
+            { "LinkedIn": "https://www.linkedin.com/in/p-uszakow/" },
+        ]
+    },
+    {
+        name: '2Długieimię Długienazwisko',
+        role: 'funkcja rola',
+        photo: author,
+        contacts: [
+            { GitHub: '#' },
+            { LinkedIn: '#' },
+        ]
+    },
+    {
+        name: '3Długieimię Długienazwisko',
+        role: 'funkcja',
+        photo: author,
+        contacts: [
+            { Figma: '#' },
+            { Facebook: '#' },
+        ]
+    },
+    {
+        name: '4Imię Nazwisko',
+        role: 'funkcja',
+        photo: author,
+        contacts: [
+            { GitHub: '#' },
+            { Facebook: '#' },
+            { Instagram: 'some link' },
+        ]
+    },
+    {
+        name: '5Imię Nazwisko',
+        role: 'funkcja',
+        photo: author,
+        contacts: [
+            { site: 'site' },
+            { 'e-mail': 'mail' },
+        ]
+    },
+    {
+        name: '6Imię Nazwisko',
+        role: 'funkcja',
+        photo: author,
+        contacts: [
+            { GitHub: '#' },
+            { LinkedIn: '#' },
+        ]
+    },
+    {
+        name: '7Imię Nazwisko',
+        role: 'funkcja w projekcie',
+        photo: author,
+        contacts: [
+            { GitHub: '#' },
+            { LinkedIn: '#' },
+        ]
+    },
+    {
+        name: '8Długieimię Długienazwisko',
+        role: 'funkcja rola',
+        photo: author,
+        contacts: [
+            { GitHub: '#' },
+            { LinkedIn: '#' },
+        ]
+    },
+    {
+        name: '9Długieimię Długienazwisko',
+        role: 'funkcja rola',
+        photo: author,
+        contacts: [
+            { GitHub: '#' },
+            { LinkedIn: '#' },
+        ]
+    },
+    {
+        name: '10Imię Nazwisko',
+        role: 'funkcja w projekcie',
+        photo: author,
+        contacts: [
+            { GitHub: '#' },
+            { LinkedIn: '#' },
+        ]
+    },
+];
+
+export default authors;
